Test client position and filename setters
Had to use a second ecco client to be able to get a previous binlog
position for the subject to start on.

I tried reusing the subject, but it seemed like the java client saved
its last position internally and started from it, so the tests passed
even if you didn’t call set_binlog_position.