Add non-blocking connect method to ecco client
…and use it in tests.

The tests don’t have to run the client in separate threads anymore,
which means that they are faster as you don’t have to sleep for a
second anymore (was used to make sure the client in the thread was
started).

The default timeout is from the tests in mysql-binlog-connector-java.

Don’t know how we can get rid of the duplicated code in TestHelper so
I’ll leave that as it is for now.