Add Rubyconverter
Related to the issue https://github.com/twingly/ecco/issues/26

We have our own (J)Ruby construct called "RowEvent" which holds
Ruby objects as well as one Java object that we receive from the
"mysql-binlog-connector-java".

"rows" that we receive is different depending on which SQL event is
fired. If "UPDATE" or "DELETE" is issued, a Java construct that uses
linked list that keeps a two-dimensional array is used. If "UPDATE"
is done, then a Java construct that uses array list that keeps a
map where the key is an array and the value is also an array.
The key holds the "value" before the update and the value holds the
"value" after the update, in the map.

The specs are now updated to use real Java objects instead of doubles
to test that the functionality of Rubyconverter works as intended,
which is to make the "row_event" uniform and only keep Ruby objects.