Refactor event listeners (#58)
Related to the issue: https://github.com/twingly/ecco/issues/25

We move all the event hashes into `event_listener.rb`. 
This way we don't have to define our events of interests in our child classes.
We add some logic to the constructor of the class that will merge the event hashes
of interest, to an instance variable. This way it is built and saved so it doesn't have to be
done again, for the same object at least. We also save the keys in an instance array, which holds
all the values that the keys point to in the hashes.
 
Close  https://github.com/twingly/ecco/issues/25