Fix README link to external codebase
In this previous commit https://github.com/twingly/ecco/commit/fb68873da42ee88e81534e46f5798ddcaaf9ddd6

I failed to see that you could actually use and create variables in markdown. This change should correct the previous commit and also make sure that we are linking to the correct page at the two different places in the README where we link to this page.